Fix instance issue
